@@ -3,10 +3,10 @@ package com.unipet7.mcommerce.activities;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.ImageView;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -14,7 +14,6 @@ import androidx.appcompat.app.ActionBar;
 import androidx.recyclerview.widget.LinearLayoutManager;
 
 import com.bumptech.glide.Glide;
-import com.google.firebase.firestore.FirebaseFirestore;
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.ProductAdapter;
 import com.unipet7.mcommerce.databinding.ActivityDetailProductBinding;
@@ -58,6 +57,7 @@ public class DetailProduct extends BaseActivity {
         }
         loadProductDetail();
         addEvents();
+        fireStoreClass.getCountUserCartItems(this);
     }
     private void loadProductDetail() {
         LoadingDialog ldDialog1 = new LoadingDialog();
@@ -67,7 +67,7 @@ public class DetailProduct extends BaseActivity {
         ldDialog1.dissmis();
     }
     public void configAdaptersProductDetail() {
-        adapter = new ProductAdapter(allProducts);
+        adapter = new ProductAdapter(allProducts, fireStoreClass);
         binding.rclProductDetail.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.HORIZONTAL, false));
         binding.rclProductDetail.setAdapter(adapter);
         binding.rclProductDetail.setHasFixedSize(true);
@@ -90,11 +90,10 @@ public class DetailProduct extends BaseActivity {
         cbFavorite.setOnCheckedChangeListener((buttonView, isChecked) -> {
             if (isChecked) {
                 // add product to favorite
-                fireStoreClass.addFavorite(this,product.getProductId());
+
                 Log.d("DetailProduct", "Add product to favorite " + product.getProductId());
             } else {
                 // remove product from favorite
-                fireStoreClass.removeFavorite(this,product.getProductId());
                 Log.d("DetailProduct", "Remove product from favorite " + product.getProductId());
             }
         });
@@ -104,6 +103,21 @@ public class DetailProduct extends BaseActivity {
             Intent intent = new Intent(DetailProduct.this, MainActivity.class);
             intent.putExtra(Constants.CART, 2);
             startActivity(intent);
+        });
+        btnAddcart.setOnClickListener(v -> {
+            fireStoreClass.getCountUserCartItems(this);
+            String productName = product.getProductname();
+            double productPrice = product.getProductprice();
+            String productImage = product.getProductImageUrl();
+            double productId = product.getProductId();
+            double numOfProduct = Double.parseDouble(binding.txtNumberOrder.getText().toString());
+            Log.d("DetailProduct", "productName: " + productName);
+            Log.d("DetailProduct", "productPrice: " + productPrice);
+            Log.d("DetailProduct", "numOfProduct: " + numOfProduct);
+            Log.d("DetailProduct", "productImageUrl: " + productImage);
+            Log.d("DetailProduct", "productId: " + productId);
+            fireStoreClass.addToCart(productId, productName, productPrice, numOfProduct, productImage);
+            Toast.makeText(DetailProduct.this, "Thêm vào giỏ hàng thành công", Toast.LENGTH_SHORT).show();
         });
     }
 
@@ -147,6 +161,7 @@ public class DetailProduct extends BaseActivity {
     }
 
     public void loadProductDetail(Product product) {
+        this.product = product;
         String name = product.getProductname();
         cbFavorite.setChecked(product.isFavorite());
         tvProductName.setText(name);
@@ -155,21 +170,7 @@ public class DetailProduct extends BaseActivity {
         String formattedPrice = String.format("%,.0f đ", price);
         double presaleprice = product.getPresaleprice();
         String formattedPreSalePrice = String.format("%,.0f đ", presaleprice);
-        btnAddcart.setOnClickListener(v -> {
-            String productName = product.getProductname();
-            double productPrice = product.getProductprice();
-            String productImage = product.getProductImageUrl();
-            double productId = product.getProductId();
-            double numOfProduct = Double.parseDouble(binding.txtNumberOrder.getText().toString());
-            String userId = fireStoreClass.getCurrentUID();
-            Log.d("DetailProduct", "productName: " + productName);
-            Log.d("DetailProduct", "productPrice: " + productPrice);
-            Log.d("DetailProduct", "numOfProduct: " + numOfProduct);
-            Log.d("DetailProduct", "productImageUrl: " + productImage);
-            Log.d("DetailProduct", "productId: " + productId);
-            fireStoreClass.addToCart(userId, productId, productName, productPrice, numOfProduct, productImage);
-            Toast.makeText(DetailProduct.this, "Thêm vào giỏ hàng thành công", Toast.LENGTH_SHORT).show();
-        });
+
         // check if product is on sale
         if (product.getSalepercent() > 0) {
             tvProductPresale.setVisibility(TextView.VISIBLE);
@@ -189,4 +190,10 @@ public class DetailProduct extends BaseActivity {
         Glide.with(this).load(product.getProductImageUrl()).into(ivProductImage);
     }
 
+    public void loadCartCount(int count) {
+        RelativeLayout cartCountLayout = binding.rlCartNumber;
+        cartCountLayout.setVisibility(RelativeLayout.VISIBLE);
+        TextView cartCount = binding.txtNumberCart;
+        cartCount.setText(String.valueOf(count));
+    }
 }
